MHER-44: Refactor into a rapidpro-csd adpater module Address CR comments
